requirements.txt
# API
fastapi
uvicorn
PyYAML
# .env
python-dotenv

# Form Data
python-multipart

requests[socks]
PySocks
stem

# Memory Storage
Redis

# SQL
SQLAlchemy
psycopg2

# Async Task
celery[redis]

# Authentication / Crypto
jwcrypto
python-jose[cryptography] # generate and verify the JWT tokens

# Hash
passlib[bcrypt] # handle password hashes

# Testing
jupyter
